app/requirements.txt
attrs==23.1.0
blinker==1.6.2
click==8.1.3
flasgger==0.9.7.1
Flask==2.3.2
Flask-SQLAlchemy==3.0.4
greenlet==2.0.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.3
mistune==3.0.1
packaging==23.1
psycopg2-binary==2.9.6
pyrsistent==0.19.3
PyYAML==6.0
six==1.16.0
SQLAlchemy==2.0.16
typing_extensions==4.6.3
Werkzeug==2.3.6
zipp==3.15.0
